requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23

# HTTP client
httpx==0.25.2
requests==2.31.0

# Machine Learning (lightweight)
# Using compatible versions that work on Windows without C++ build tools
scikit-learn==1.3.2; platform_system != "Windows"
scikit-learn==1.3.2; platform_system == "Windows" and python_version >= "3.11"
numpy==1.26.2
pandas==2.1.3

# Time series (optional, for Prophet alternative)
# prophet==1.1.5  # Uncomment if needed

# Logging
structlog==23.2.0

# Configuration
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For async testing

# Utilities
python-dateutil==2.8.2
python-json-logger==2.0.7

